refactor: solution cleanup.
