+++ src/Imlight.Engine/Processor.cs
