+++ src/Imlight.Engine/ProcessorManager.cs
