+++ src/Imlight.Engine/WizardMessageContext.cs
