+++ src/Imlight.Realm/RealmManager.cs
