--- src/Login/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Login
-{
-    public class Class1
-    {
-    }
-}
