--- src/Processor/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Processor
-{
-    public class Class1
-    {
-    }
-}
