--- src/Storage/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Storage
-{
-    public class Class1
-    {
-    }
-}
