--- src/StorageBackup/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace StorageBackup
-{
-    public class Class1
-    {
-    }
-}
